+++ src/main/java/hu/oe/nik/szfmv17t/automatedcar/Driver.java
@@ -1,7 +1,7 @@
-package hu.oe.nik.automatedcar;
+package hu.oe.nik.szfmv17t.automatedcar;
 
-import hu.oe.nik.automatedcar.bus.Signal;
-import hu.oe.nik.automatedcar.bus.VirtualFunctionBus;
+import hu.oe.nik.szfmv17t.automatedcar.bus.Signal;
+import hu.oe.nik.szfmv17t.automatedcar.bus.VirtualFunctionBus;
 
 /**
  * This is an example class for the SystemComponent.
