+++ src/main/java/hu/oe/nik/szfmv17t/automatedcar/SystemComponent.java
@@ -1,6 +1,6 @@
-package hu.oe.nik.automatedcar;
+package hu.oe.nik.szfmv17t.automatedcar;
 
-import hu.oe.nik.automatedcar.bus.VirtualFunctionBus;
+import hu.oe.nik.szfmv17t.automatedcar.bus.VirtualFunctionBus;
 
 /**
  *
