+++ src/main/java/hu/oe/nik/szfmv17t/automatedcar/bus/Bus.java
@@ -1,4 +1,4 @@
-package hu.oe.nik.automatedcar.bus;
+package hu.oe.nik.szfmv17t.automatedcar.bus;
 
 public class Bus {
 	/* fields */
