+++ src/main/java/hu/oe/nik/szfmv17t/automatedcar/bus/Signal.java
@@ -1,4 +1,4 @@
-package hu.oe.nik.automatedcar.bus;
+package hu.oe.nik.szfmv17t.automatedcar.bus;
 
 /**
  * This class represent a single signal on the bus.
