+++ src/main/java/hu/oe/nik/szfmv17t/automatedcar/powertrainsystem/PowertrainSystem.java
@@ -1,7 +1,7 @@
-package hu.oe.nik.automatedcar.powertrainsystem;
+package hu.oe.nik.szfmv17t.automatedcar.powertrainsystem;
 
-import hu.oe.nik.automatedcar.SystemComponent;
-import hu.oe.nik.automatedcar.bus.Signal;
+import hu.oe.nik.szfmv17t.automatedcar.SystemComponent;
+import hu.oe.nik.szfmv17t.automatedcar.bus.Signal;
 
 public class PowertrainSystem extends SystemComponent {
