+++ src/test/java/hu/oe/nik/szfmv17t/automatedcar/bus/BusTest.java
@@ -1,4 +1,4 @@
-package hu.oe.nik.automatedcar.bus;
+package hu.oe.nik.szfmv17t.automatedcar.bus;
 
 import static org.junit.Assert.assertEquals;
